+++ packages/dodo-ui/src/lib/index.ts
@@ -44,28 +44,6 @@ export {
 } from './components/Form/FormField/FormField.svelte';
 
 /**
- * Checkbox component and related prop types.
- */
-export {
-  default as Checkbox,
-  type CheckboxProps,
-} from './components/Form/Checkbox/Checkbox.svelte';
-
-/**
- * RadioGroup component and related prop types.
- */
-export {
-  default as RadioGroup,
-  type RadioGroupProps,
-} from './components/Form/RadioGroup/RadioGroup.svelte';
-export { default as Radio, type RadioProps } from './components/Form/RadioGroup/Radio/Radio.svelte';
-
-/**
- * Switch component and related prop types.
- */
-export { default as Switch, type SwitchProps } from './components/Form/Switch/Switch.svelte';
-
-/**
  * UtilityButton component and related prop types.
  */
 export {
@@ -74,19 +52,6 @@ export {
   type UtilityButtonAsAnchorProps,
   type UtilityButtonProps,
 } from './components/Form/UtilityButton/UtilityButton.svelte';
-
-/**
- * Select component and related prop types.
- */
-export {
-  default as Select,
-  type SelectProps,
-  type SelectOption,
-} from './components/Form/Select/Select.svelte';
-
-export { type SelectInputProps } from './components/Form/Select/SelectInput.svelte';
-
-export { type SelectPopupProps } from './components/Form/Select/SelectPopup.svelte';
 
 // =====================================================
 // Layout Components – Public API